requirements.txt
Flask
python-dotenv
requests
Pillow
openai
gunicorn
pillow-avif-plugin